add search-app flag to dictionary connector schema
Cloud Search recently added a new field to the dictionary entry well-known schema. This field, "_onlyApplicableForAttachedSearchApplications", tells Cloud Search that the synonym item should only be applied to search applications that include the synonym's data source. This contrasts with the default behavior, which is that a synonym applies to all search applications at the domain.

Since the dictionary connector relies on an explicitly defined schema, this field must be defined here to enable its use with the sample dictionary connector.